CXP Agregado

+++ CapaDatos/proc_CargarTodasLineasCreditoVentas_Result.cs
@@ -15,10 +15,10 @@ namespace CapaDatos
     {
         public int LineaCreditoVentaID { get; set; }
         public System.DateTime Fecha { get; set; }
-        public Nullable<double> MontoFactura { get; set; }
+        public Nullable<decimal> MontoFactura { get; set; }
         public string Cliente { get; set; }
         public int Factura { get; set; }
-        public Nullable<double> BalancePendiente { get; set; }
+        public Nullable<decimal> BalancePendiente { get; set; }
         public bool Completado { get; set; }
     }
 }
